Prevent -Wsign-compare warnings in public htslib/*.h headers
(Also fix nearby htslib/hfile.h whitespace.)

+++ htslib/hts_endian.h
@@ -242,7 +242,7 @@ static inline int16_t le_to_i16(const uint8_t *buf) {
  */
 static inline int32_t le_to_i32(const uint8_t *buf) {
     uint32_t v = le_to_u32(buf);
-    return v < 0x80000000U ? v : -((int32_t) (0xffffffffU - v)) - 1;
+    return v < 0x80000000U ? (int32_t) v : -((int32_t) (0xffffffffU - v)) - 1;
 }
 
 /// Get an int64_t value from an unsigned byte array
@@ -254,7 +254,7 @@ static inline int32_t le_to_i32(const uint8_t *buf) {
 static inline int64_t le_to_i64(const uint8_t *buf) {
     uint64_t v = le_to_u64(buf);
     return (v < 0x8000000000000000ULL
-            ? v : -((int64_t) (0xffffffffffffffffULL - v)) - 1);
+            ? (int64_t) v : -((int64_t) (0xffffffffffffffffULL - v)) - 1);
 }
 
 // Converting the other way is easier as signed -> unsigned is well defined.
